+++ lib/src/expression/group.dart
@@ -30,7 +30,7 @@ class ExpressionGroup<T> {
   /// delimiter, the `value` and `right` delimiter.
   void wrapper<L, R>(Parser<L> left, Parser<R> right,
           T Function(L left, T value, R right) callback) =>
-      _wrapper.add(seq3(left, _loopback, right).map3(callback));
+      _wrapper.add((left, _loopback, right).toSequenceParser().map3(callback));
 
   Parser<T> _buildWrapper(Parser<T> inner) => buildChoice([..._wrapper, inner]);
 
@@ -44,8 +44,9 @@ class ExpressionGroup<T> {
 
   Parser<T> _buildPrefix(Parser<T> inner) => _prefix.isEmpty
       ? inner
-      : seq2(buildChoice(_prefix).star(), inner).map2((prefix, value) =>
-          prefix.reversed.fold(value, (each, result) => result.call(each)));
+      : (buildChoice(_prefix).star(), inner).toSequenceParser().map2(
+          (prefix, value) =>
+              prefix.reversed.fold(value, (each, result) => result.call(each)));
 
   final List<Parser<ExpressionResultPrefix<T, void>>> _prefix = [];
 
@@ -57,8 +58,9 @@ class ExpressionGroup<T> {
 
   Parser<T> _buildPostfix(Parser<T> inner) => _postfix.isEmpty
       ? inner
-      : seq2(inner, buildChoice(_postfix).star()).map2((value, postfix) =>
-          postfix.fold(value, (each, result) => result.call(each)));
+      : (inner, buildChoice(_postfix).star()).toSequenceParser().map2(
+          (value, postfix) =>
+              postfix.fold(value, (each, result) => result.call(each)));
 
   final List<Parser<ExpressionResultPostfix<T, void>>> _postfix = [];