+++ lib/src/parser/action/flatten.dart
@@ -33,19 +33,18 @@ class FlattenParser<R> extends DelegateParser<R, String> {
     // If we have a message we can switch to fast mode.
     if (message != null) {
       final position = delegate.fastParseOn(context.buffer, context.position);
-      if (position < 0) {
-        return context.failure(message!);
-      }
+      if (position < 0) return context.failure(message!);
       final output = context.buffer.substring(context.position, position);
       return context.success(output, position);
     } else {
       final result = delegate.parseOn(context);
-      if (result.isSuccess) {
-        final output =
-            context.buffer.substring(context.position, result.position);
-        return result.success(output);
+      switch (result) {
+        case Success(position: final position):
+          final output = context.buffer.substring(context.position, position);
+          return result.success(output);
+        case Failure(message: final message):
+          return result.failure(message);
       }
-      return result.failure(result.message);
     }
   }