+++ lib/src/parser/action/where.dart
@@ -66,10 +66,11 @@ class WhereParser<R> extends DelegateParser<R, R> {
   @override
   Result<R> parseOn(Context context) {
     final result = delegate.parseOn(context);
-    if (result is Success<R> && !predicate(result.value)) {
-      return failureFactory(context, result);
-    }
-    return result;
+    return switch (result) {
+      Success(value: final value) when !predicate(value) =>
+        failureFactory(context, result),
+      _ => result
+    };
   }
 
   @override